+++ packages/cli/src/ui/constants.ts
@@ -12,22 +12,4 @@ export const BOX_PADDING_X = 1;
 export const UI_WIDTH =
   EstimatedArtWidth + BOX_PADDING_X * 2 + BoxBorderWidth * 2; // ~63
 
-export const WITTY_LOADING_PHRASES = [
-  'Consulting the digital spirits...',
-  'Reticulating splines...',
-  'Warming up the AI hamsters...',
-  'Asking the magic conch shell...',
-  'Generating witty retort...',
-  'Polishing the algorithms...',
-  "Don't rush perfection (or my code)...",
-  'Brewing fresh bytes...',
-  'Counting electrons...',
-  'Engaging cognitive processors...',
-  'Checking for syntax errors in the universe...',
-  'One moment, optimizing humor...',
-  'Shuffling punchlines...',
-  'Untangling neural nets...',
-  'Compiling brilliance...',
-];
-export const PHRASE_CHANGE_INTERVAL_MS = 15000;
 export const STREAM_DEBOUNCE_MS = 100;
